+++ CHANGELOG.md
@@ -3,6 +3,9 @@ Next
 
 ## New features
 
+- New option --print-srcaddr to show used IP source address in output (#461, thanks @gsnw-sebast)
+- Add IPv6 support to --print-srcaddr (#462, thanks @gsnw-sebast)
+
 ## Bugfixes and other changes
 
 - ci: Removed travis-ci (#446, thanks @gsnw-sebast)